+++ RELEASES.md
@@ -22,6 +22,7 @@ Upcoming
 - Add attribute version to JsonPipe
 - Add dynamic throttling in senders
 - Add property jdbc.convertFieldnamesToUppercase
+- Add attribute hideRegex to pipes (next to global property)
 - Fix UUIDGeneratorPipe when stubbed remove attribute type